# swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Meu Amigo PET",
		"version": "1.0.0",
		"description": "Meu Amigo PET é uma API REST cujo objetivo principal é conectar usuários, sejam eles pais de PETs ou aspirantes ao cargo, com objetivos semelhantes: adoção ou cruzamento. A conexão entre usuários se dá por meio de anúncios, os quais podem ser de adoção ou cruzamento. Assim, usuários logados podem acessar o 'feed de anúnicos' e obter informações de pets e usuários. A API visa atender usuários comuns e administradores com funcionalidades diferentes para cada um dos tipos de contas. Os usuários devem estar logados no servidor para acessar as diferentes rotas e suas funcionalidades. A aplicação foi construida usando-se Node.JS, módulos NPM (específicados no arquivo **package.json**), o framework Express e o [GitHub](https: //github.com/Alisson3366/projeto-de-back-end)."
	},
	"servers": [
		{
			"url": "http://localhost:8080"
		}
	],
...
	"paths": {
...
		"/pets": {
			"get": {
				"tags": ["pets"],
				"summary": "Consulta de todos os pets",
				"description": "Consulta de todos os pets",
				"operationId": "consultaPets",
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PETS"
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/pet/{id}": {
			"get": {
				"tags": ["pets"],
				"summary": "Consulta de pet específico pelo ID",
				"description": "Consulta de pet específico pelo ID",
				"operationId": "consultaPetId",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do pet a ser consultado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PETS"
								}
							}
						}
					},
					"400": {
						"description": "ID de pet inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Pet não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/pets/usuario": {
			"get": {
				"tags": ["pets"],
				"summary": "Consulta dos pets do usuário logado",
				"description": "Consulta dos pets do usuário logado",
				"operationId": "consultaPetsUsuario",
				"responses": {
					"200": {
						"description": "Consulta efetuada com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PETS"
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["pets"],
				"summary": "Cria novo pet do usuário logado",
				"description": "Cria novo pet do usuário logado",
				"operationId": "adicionaPetUsuario",
				"requestBody": {
					"description": "Cria novo pet do usuário logado",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"nome": {
										"type": "string",
										"example": "PITOCO"
									},
									"sexo": {
										"type": "string",
										"example": "fêmea"
									},
									"raca": {
										"type": "string",
										"example": "srd"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Pet adicionado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Nome, sexo ou raca inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/pet/usuario/{id}": {
			"patch": {
				"tags": ["pets"],
				"summary": "Atualiza um pet já existente de um usuário",
				"description": "Atualiza um pet já existente de um usuário",
				"operationId": "atualizaPetUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do pet a ser atualizado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"requestBody": {
					"description": "Atualiza um pet já existente de um usuário",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"nome": {
										"type": "string",
										"example": "PITOCO"
									},
									"sexo": {
										"type": "string",
										"example": "fêmea"
									},
									"raca": {
										"type": "string",
										"example": "srd"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Pet atualizado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de pet inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Pet não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Nome, sexo ou raca inválidos!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["pets"],
				"summary": "Deleta o pet de um usuário da API",
				"description": "Deleta o pet de um usuário da API",
				"operationId": "deletaPetUsuario",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "ID do pet a ser deletado.",
						"required": true,
						"schema": {
							"type": "string",
							"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Pet deletado com sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "ID de pet inválido!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Token não informado ou inválido. Acesso negado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Pet não localizado!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno na aplicação!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Erro": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"PETS": {
				"required": ["nome", "sexo", "raca"],
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"donoPet": {
						"type": "string",
						"example": "3732f525-2632-4460-b67d-fcad2262e8f1"
					},
					"nome": {
						"type": "string",
						"example": "NINA"
					},
					"sexo": {
						"type": "string",
						"example": "fêmea"
					},
					"raca": {
						"type": "string",
						"example": "beagle"
					}
				}
			}
		}
	}
}